refacto: create component for focused-tooltip
