+++ mon-pix/app/components/challenge-statement.js
@@ -1,8 +1,8 @@
+import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
 import config from 'mon-pix/config/environment';
-import Component from '@glimmer/component';
 import sortBy from 'lodash/sortBy';
 import ENV from 'mon-pix/config/environment';
 
@@ -10,17 +10,13 @@ const PREFERRED_ATTACHMENT_FORMATS = ['docx', 'xlsx', 'pptx'];
 
 export default class ChallengeStatement extends Component {
   @service mailGenerator;
-  @service currentUser;
   @service intl;
 
   @tracked selectedAttachmentUrl;
   @tracked displayAlternativeInstruction = false;
-  @tracked shouldDisplayTooltip = false;
-
   constructor() {
     super(...arguments);
     this._initialiseDefaultAttachment();
-    this._showTooltip();
   }
 
   get isFocusedChallengeToggleEnabled() {
@@ -55,48 +51,6 @@ export default class ChallengeStatement extends Component {
 
   get id() {
     return 'challenge_statement_' + this.args.challenge.id;
-  }
-
-  get isFocusedChallenge() {
-    return this.args.challenge.focused;
-  }
-
-  @action
-  hideTooltip() {
-    this.shouldDisplayTooltip = false;
-    this._notifyChallengeTooltipIsClosed();
-  }
-
-  _showTooltip() {
-    if (this.isFocusedChallenge && this.currentUser.user && !this.currentUser.user.hasSeenFocusedChallengeTooltip) {
-      this.shouldDisplayTooltip = true;
-    }
-    else if (this.isFocusedChallenge && this.currentUser.user && this.currentUser.user.hasSeenFocusedChallengeTooltip) {
-      this._notifyChallengeTooltipIsClosed();
-    }
-    else if (this.isFocusedChallenge) {
-      this._notifyChallengeTooltipIsClosed();
-    }
-  }
-
-  @action
-  showTooltipOnMouseEnter() {
-    if (this.isFocusedChallenge && this.currentUser.user.hasSeenFocusedChallengeTooltip) {
-      this.shouldDisplayTooltip = true;
-    }
-  }
-
-  @action
-  hideTooltipOnMouseLeave() {
-    this.shouldDisplayTooltip = false;
-  }
-
-  get shouldDisplayButton() {
-    return this.isFocusedChallenge && !this.currentUser.user.hasSeenFocusedChallengeTooltip;
-  }
-
-  _notifyChallengeTooltipIsClosed() {
-    this.args.onTooltipClose();
   }
 
   @action
