+++ mon-pix/app/components/challenge/challenge-statement/focused-tooltip.hbs
@@ -1,0 +1,23 @@
+<div class="challenge-statement-instruction__tag challenge-statement-instruction__tag{{if this.isFocusedChallenge "--focused" "--regular"}}"
+  {{on 'mouseenter' this.showTooltipOnMouseEnter}}
+  {{on 'mouseleave' this.hideTooltipOnMouseLeave}}>
+  {{#if this.shouldDisplayTooltip}}
+    <div role="tooltip"
+         id="challenge-statement-tag--tooltip"
+         class="challenge-statement__tag-information"
+      {{on-key 'Escape' this.hideTooltip event='keyup'}}>
+      {{#if this.isFocusedChallenge}}
+        <div class="challenge-statement-tag-information__text">
+          {{t 'pages.challenge.statement.tooltip.focused'}}
+        </div>
+      {{/if}}
+      {{#if this.shouldDisplayButton}}
+        <PixButton class="challenge-statement-tag-information__button"
+                   @size="small"
+                   @triggerAction={{this.hideTooltip}}>
+          {{t 'pages.challenge.statement.tooltip.close'}}
+        </PixButton>
+      {{/if}}
+    </div>
+  {{/if}}
+</div>
